infrastructure/lambda/upload/requirements.txt
cryptography>=44.0.1
python-jose[cryptography]>=3.3.0
requests==2.32.3